Consolidate pdf connectors
